scripts/requirements.txt
numpy
xarray
zarr
dask
pynrrd
numba
allensdk
pandas
psycopg2
h5py
h5netcdf
toolz
requests
lmdb
future
wget
